Update SyncTest sample for clarity
Both input and output textures are shown in a split manner. So, if sync is not working or Unreal's framerate is higher than TouchEngine's, a horizontal tear will be seen. And each texture has an ID number to better distinguish them from each other.

TouchEngine's frame rate is 121 for the SyncTest Blueprint to ensure higher frequency than the Unreal editor's default max of 120 fps. This guarantees a frame update in TouchEngine for each Unreal frame.